@@ -5,10 +5,12 @@
   "main": "src/index.js",
   "scripts": {
     "start": "webpack-dev-server --open --mode development",
+    "start-test": "NODE_ENV=test node app.js",
     "start-prod": "node app.js",
     "test": "jest",
     "eslint": "eslint './**/*.{js,jsx}'",
-    "build": "webpack --mode production"
+    "build": "webpack --mode production",
+    "test:e2e": "npx cypress run"
   },
   "repository": {
     "type": "git",
@@ -47,8 +49,8 @@
   "dependencies": {
     "axios": "^0.19.2",
     "express": "^4.17.1",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
     "react-router-dom": "^5.2.0"
   }
 }
